requirements.txt
pymongo~=3.10.1
dnspython~=1.16.0